+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Expose session timeout. #887
+- Added `event.origin` and `event.environment` tags to determine where events originate from. #890
 
 ## 1.4.1
 
@@ -134,7 +135,7 @@ New way to import and init the SDK:
 import * as Sentry from "@sentry/react-native";
 
 Sentry.init({
-  dsn: "DSN",
+  dsn: "DSN"
 });
 ```
 
@@ -370,7 +371,7 @@ To activate it set:
 
 ```js
 Sentry.config("___DSN___", {
-  deactivateStacktraceMerging: false,
+  deactivateStacktraceMerging: false
 });
 ```